fix map! working with sliders!
