+++ helpers.js
@@ -1,3 +1,5 @@
+import slugify from "slugify";
+
 export function formatAndScaleValue(value, oblastKey, oblastScales) {
   // Get the proper scale for this oblast name
   const scale = oblastScales ? oblastScales[oblastKey] : 1;
@@ -11,4 +13,36 @@ export function formatAndScaleValue(value, oblastKey, oblastScales) {
 
 export function formatValue(value) {
   return (+value).toFixed(1);
+}
+
+// We use the spelling from the data set - NOT the map property data
+function standardizeOblastSpelling(oblastName) {
+  switch (oblastName) {
+    // TODO crimea / sevastopal are missing due to russian occupation.
+    case "khmelnytskyy":
+      return "khmelnytskiy";
+    case "kiev":
+      return "kyiv";
+    case "kiev-city":
+      return "kyiv";
+    case "odessa":
+      return "odesa";
+    case "mykolayiv":
+      return "mikolayiv";
+    case "transcarpathia":
+      return "zakarpattya";
+    default:
+      return oblastName
+  }
+}
+
+// Normalize our oblast name using slugify
+export function normalizeOblastName(key) {
+  if (!key) return key;
+  return standardizeOblastSpelling(
+    slugify(key, {
+      strict: true,
+      lower: true,
+    })
+  );
 }
